+++ src/components/ui/form/index.js
@@ -1,5 +1,20 @@
+import CheckboxUi from '@/components/ui/form/CheckboxUi/CheckboxUi'
+import FileUi from '@/components/ui/form/FileUi/FileUi'
 import FormField from '@/components/ui/form/FormField/FormField'
+import FormUi from '@/components/ui/form/FormUi/FormUi'
+import InputUi from '@/components/ui/form/InputUi/InputUi'
+import RadioUi from '@/components/ui/form/RadioUi/RadioUi'
+import SelectUi from '@/components/ui/form/SelectUi/SelectUi'
+import TexareaUi from '@/components/ui/form/TexareaUi/TexareaUi'
+
 
 export {
+  CheckboxUi,
+  FileUi,
   FormField,
+  FormUi,
+  InputUi,
+  RadioUi,
+  SelectUi,
+  TexareaUi
 }