+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - RadioUi,
   - SelectUi,
   - TexareaUi
+  - FieldUi
+  - FormUi
+- Add form builder
 ### Changed
 ### Deprecated
 ### Removed