+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - FieldUi
   - FormUi
 - Add form builder
+- Updated Vue from 2.5.2 to 2.5.13
 ### Changed
 ### Deprecated
 ### Removed